revert changes to lastReceivedMerkleIndex

+++ beacon-chain/powchain/log_processing.go
@@ -109,21 +109,21 @@ func (s *Service) ProcessDepositLog(ctx context.Context, depositLog gethTypes.Lo
 	// ETH1.0 network, and prevents us from updating our trie
 	// with the same log twice, causing an inconsistent state root.
 	index := binary.LittleEndian.Uint64(merkleTreeIndex)
-	if index <= s.lastReceivedMerkleIndex {
+	if int64(index) <= s.lastReceivedMerkleIndex {
 		return nil
 	}
 
-	if index != s.lastReceivedMerkleIndex+1 {
+	if int64(index) != s.lastReceivedMerkleIndex+1 {
 		missedDepositLogsCount.Inc()
 		if s.requestingOldLogs {
 			return errors.New("received incorrect merkle index")
 		}
-		if err := s.requestMissingLogs(ctx, depositLog.BlockNumber, index-1); err != nil {
+		if err := s.requestMissingLogs(ctx, depositLog.BlockNumber, int64(index-1)); err != nil {
 			return errors.Wrap(err, "could not get correct merkle index")
 		}
 
 	}
-	s.lastReceivedMerkleIndex = index
+	s.lastReceivedMerkleIndex = int64(index)
 
 	// We then decode the deposit input in order to create a deposit object
 	// we can store in our persistent DB.
@@ -285,7 +285,7 @@ func (s *Service) processPastLogs(ctx context.Context) error {
 	}
 	for currentBlockNum < latestFollowHeight {
 		// stop requesting, if we have all the logs
-		if logCount == s.lastReceivedMerkleIndex+1 {
+		if logCount == uint64(s.lastReceivedMerkleIndex+1) {
 			break
 		}
 		start := currentBlockNum
@@ -300,7 +300,7 @@ func (s *Service) processPastLogs(ctx context.Context) error {
 			FromBlock: big.NewInt(int64(start)),
 			ToBlock:   big.NewInt(int64(end)),
 		}
-		remainingLogs := logCount - s.lastReceivedMerkleIndex + 1
+		remainingLogs := logCount - uint64(s.lastReceivedMerkleIndex+1)
 		// only change the end block if the remaining logs are below the required log limit.
 		if remainingLogs < depositlogRequestLimit && end >= latestFollowHeight {
 			query.ToBlock = big.NewInt(int64(latestFollowHeight))
@@ -371,7 +371,7 @@ func (s *Service) requestBatchedLogs(ctx context.Context) error {
 
 // requestMissingLogs requests any logs that were missed by requesting from previous blocks
 // until the current block(exclusive).
-func (s *Service) requestMissingLogs(ctx context.Context, blkNumber uint64, wantedIndex uint64) error {
+func (s *Service) requestMissingLogs(ctx context.Context, blkNumber uint64, wantedIndex int64) error {
 	// Prevent this method from being called recursively
 	s.requestingOldLogs = true
 	defer func() {
